Split connector handler into multiple functions
This commit splits the old connector process function into multiple structs with separate process functions. This should make the code a bit more organized instead of everything just being one big function. All connector request structs now implement a `Process` trait that processes the request and returns a connector response. Structs are added to the connector request enum using the `build_request_enum` macro.

Additionally, error handling for the connector has been improved and most of the relevant objects have been documented. This is one of the steps to make the connector API more stable and easier to use in other applications.

+++ native/src/console/site.rs
@@ -221,9 +221,11 @@ impl Run for SiteUninstallCommand {
             .retain(|id| *id != self.id);
         let site = storage.sites.remove(&self.id);
 
-        if let Some(site) = site {
-            site.uninstall_system_integration(&dirs)
-                .context("Failed to uninstall system integration")?;
+        if self.system_integration {
+            if let Some(site) = site {
+                site.uninstall_system_integration(&dirs)
+                    .context("Failed to uninstall system integration")?;
+            }
         }
 
         storage.write(&dirs)?;
@@ -252,8 +254,7 @@ impl Run for SiteUpdateCommand {
         }
 
         if self.system_integration {
-            site.install_system_integration(&dirs)
-                .context("Failed to update system integration")?;
+            site.update_system_integration(&dirs).context("Failed to update system integration")?;
         }
 
         storage.write(&dirs)?;
